requirements.txt
asgiref==3.8.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.7
django-allauth==0.63.6
django-environ==0.11.2
django-filter==24.3
django-htmx==1.19.0
django-model-utils==4.5.1
django-storages==1.14.4
djangorestframework==3.15.2
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
idna==3.7
jwt==1.3.1
Markdown==3.6
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
requests==2.32.3
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
Unipath==1.1
urllib3==2.2.2